% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dse_gen_simple_app_r.R
\name{dse_gen_simple_app_r}
\alias{dse_gen_simple_app_r}
\title{generateur d une application shiny simple au format app.R
dans un repertoire}
\usage{
dse_gen_simple_app_r(chemin)
}
\arguments{
\item{chemin}{une chaine de caracteres qui indique ou construire l application}
}
\value{
la fonction construit un repertoire qui contient l'application
}
\description{
generateur d une application shiny simple au format app.R
dans un repertoire
}
