% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dse_ui_regexp.R
\name{dse_ui_regexp}
\alias{dse_ui_regexp}
\title{traduction grille bootstrap vers dsfr}
\usage{
dse_ui_regexp(ui)
}
\arguments{
\item{ui}{une liste ui générée par shiny::fluidpage}
}
\value{
la fonction retourne une ui interprétable par dsfr
}
\description{
traduction grille bootstrap vers dsfr
}
